Refactor Data class to return Game, Team, and GameTeam data array
